// src/knowledge/symbols/models/SymbolKind.ts
export enum SymbolKind {
  Module = 'Module',
  Class = 'Class',
  Interface = 'Interface',
  Function = 'Function',
  Method = 'Method',
  Variable = 'Variable',
  Constant = 'Constant',
  Route = 'Route',
  Middleware = 'Middleware',
  Endpoint = 'Endpoint',
  Entity = 'Entity',
  Config = 'Config',
  Unknown = 'Unknown'
}